refactor(driver): move test context detection helpers into test_context module
Extracted functions related to identifying test-only code into a new module
`test_context`. This includes detection of test modules by name, test functions marked
by harness descriptors or test attributes, and handling companion modules from test
framework macros like `rstest`. The main driver module imports these helpers instead
of defining them inline, improving code organization and maintainability.

Also updated tests and imports accordingly to use the new module.
